shim: organize code to different modules
this commit organized code to different modules:
- constants: to handling constants across entire crate
- engine: the execution engine
- source: manages application sources
- stdio_hook: a trigger hook for stdio
- trigger: manages spin triggers
- utils: utility functions

it also removes ResolvedAppSource as it's not needed
